+++ Ruby Rose/Modules/Moderation/Whitelist.cs
@@ -5,10 +5,10 @@ using System.Threading.Tasks;
 using Discord;
 using Discord.Commands;
 using MongoDB.Driver;
-using RubyRose.Common;
 using RubyRose.Common.Preconditions;
 using RubyRose.Database;
 using NLog;
+using RubyRose.Database.Models;
 
 namespace RubyRose.Modules.Moderation
 {
@@ -18,15 +18,11 @@ namespace RubyRose.Modules.Moderation
         [Group("Whitelist"), Name("Whitelist")]
         public class WhitelistCommands : ModuleBase
         {
-            private static Logger logger = LogManager.GetCurrentClassLogger();
-            private readonly CommandService _service;
-            private readonly IDependencyMap _map;
+            private static readonly Logger Logger = LogManager.GetCurrentClassLogger();
             private readonly MongoClient _mongo;
 
-            public WhitelistCommands(CommandService service, IDependencyMap map)
+            public WhitelistCommands(IDependencyMap map)
             {
-                _service = service ?? throw new ArgumentNullException(nameof(service));
-                _map = map;
                 _mongo = map.Get<MongoClient>();
             }
 
@@ -36,35 +32,38 @@ namespace RubyRose.Modules.Moderation
             public async Task Add(CommandInfo command, ITextChannel channel = null)
             {
                 channel = channel ?? Context.Channel as ITextChannel;
-                var newWhite = new Whitelists
+                if (channel != null)
                 {
-                    GuildId = Context.Guild.Id,
-                    ChannelId = channel.Id,
-                    Name = command.Name
-                };
-
-                var allWhitelists = _mongo.GetCollection<Whitelists>(Context.Client);
-                var whitelists = await GetCommandWhitelists(allWhitelists, Context.Guild, command.Name);
-
-                if (whitelists != null)
-                {
-                    if (!whitelists.Exists(c => c.ChannelId == channel.Id))
+                    var newWhite = new Whitelists
                     {
-                        await allWhitelists.InsertOneAsync(newWhite);
-                        await ReplyAsync($"Command `{command.Name}` is now Whitelisted to `{channel.Name}`");
-                        logger.Info($"Command {command.Name} now whitelisted to {channel.Name} on {Context.Guild.Name}");
+                        GuildId = Context.Guild.Id,
+                        ChannelId = channel.Id,
+                        Name = command.Name
+                    };
+
+                    var allWhitelists = _mongo.GetCollection<Whitelists>(Context.Client);
+                    var whitelists = await GetCommandWhitelists(allWhitelists, Context.Guild, command.Name);
+
+                    if (whitelists != null)
+                    {
+                        if (!whitelists.Exists(c => c.ChannelId == channel.Id))
+                        {
+                            await allWhitelists.InsertOneAsync(newWhite);
+                            await ReplyAsync($"Command `{command.Name}` is now Whitelisted to `{channel.Name}`");
+                            Logger.Info($"Command {command.Name} now whitelisted to {channel.Name} on {Context.Guild.Name}");
+                        }
+                        else
+                        {
+                            await ReplyAsync($"Command {command.Name} already whitelisted to `{channel.Name}`");
+                            Logger.Warn($"Failed to Whitelist {command.Name} to {channel.Name} on {Context.Guild.Name}, already Whitelisted");
+                        }
                     }
                     else
                     {
-                        await ReplyAsync($"Command {command.Name} already whitelisted to `{channel.Name}`");
-                        logger.Warn($"Failed to Whitelist {command.Name} to {channel.Name} on {Context.Guild.Name}, already Whitelisted");
+                        await allWhitelists.InsertOneAsync(newWhite);
+                        await ReplyAsync($"Command `{command.Name}` is now Whitelisted to `{channel.Name}`");
+                        Logger.Info($"Command {command.Name} now whitelisted to {channel.Name} on {Context.Guild.Name}");
                     }
-                }
-                else
-                {
-                    await allWhitelists.InsertOneAsync(newWhite);
-                    await ReplyAsync($"Command `{command.Name}` is now Whitelisted to `{channel.Name}`");
-                    logger.Info($"Command {command.Name} now whitelisted to {channel.Name} on {Context.Guild.Name}");
                 }
             }
 
@@ -79,23 +78,23 @@ namespace RubyRose.Modules.Moderation
 
                 if (whitelists != null)
                 {
-                    if (whitelists.Exists(c => c.ChannelId == channel.Id))
+                    if (whitelists.Exists(c => channel != null && c.ChannelId == channel.Id))
                     {
-                        var whitelist = whitelists.First(c => c.ChannelId == channel.Id);
+                        var whitelist = whitelists.First(c => channel != null && c.ChannelId == channel.Id);
                         await allWhitelists.DeleteAsync(whitelist);
-                        await ReplyAsync($"Command `{command.Name}` is no longer Whitelisted to `{channel.Name}`");
-                        logger.Info($"Command {command.Name} no longer whitelisted to {channel.Name} on {Context.Guild.Name}");
+                        await ReplyAsync($"Command `{command.Name}` is no longer Whitelisted to `{channel?.Name}`");
+                        Logger.Info($"Command {command.Name} no longer whitelisted to {channel?.Name} on {Context.Guild.Name}");
                     }
                     else
                     {
-                        await ReplyAsync($"Command {command.Name} not whitelisted to `{channel.Name}`");
-                        logger.Warn($"Failed to remove Whitelist {command.Name} from {channel.Name} on {Context.Guild.Name}, not Whitelisted");
+                        await ReplyAsync($"Command {command.Name} not whitelisted to `{channel?.Name}`");
+                        Logger.Warn($"Failed to remove Whitelist {command.Name} from {channel?.Name} on {Context.Guild.Name}, not Whitelisted");
                     }
                 }
                 else
                 {
-                    await ReplyAsync($"Command {command.Name} not whitelisted to `{channel.Name}`");
-                    logger.Warn($"Failed to remove Whitelist {command.Name} from {channel.Name} on {Context.Guild.Name}, not Whitelisted");
+                    await ReplyAsync($"Command {command.Name} not whitelisted to `{channel?.Name}`");
+                    Logger.Warn($"Failed to remove Whitelist {command.Name} from {channel?.Name} on {Context.Guild.Name}, not Whitelisted");
                 }
             }
 
@@ -105,35 +104,38 @@ namespace RubyRose.Modules.Moderation
             public async Task All(ITextChannel channel = null)
             {
                 channel = channel ?? Context.Channel as ITextChannel;
-                var newWhite = new Whitelists
+                if (channel != null)
                 {
-                    GuildId = Context.Guild.Id,
-                    ChannelId = channel.Id,
-                    Name = "all"
-                };
-
-                var allWhitelists = _mongo.GetCollection<Whitelists>(Context.Client);
-                var whitelists = await GetCommandWhitelists(allWhitelists, Context.Guild, "all");
-
-                if (whitelists != null)
-                {
-                    if (!whitelists.Exists(c => c.ChannelId == channel.Id))
+                    var newWhite = new Whitelists
                     {
-                        await allWhitelists.InsertOneAsync(newWhite);
-                        await ReplyAsync($"All Commands are now Whitelisted to `{channel.Name}`");
-                        logger.Info($"All Commands now whitelisted to {channel.Name} on {Context.Guild.Name}");
+                        GuildId = Context.Guild.Id,
+                        ChannelId = channel.Id,
+                        Name = "all"
+                    };
+
+                    var allWhitelists = _mongo.GetCollection<Whitelists>(Context.Client);
+                    var whitelists = await GetCommandWhitelists(allWhitelists, Context.Guild, "all");
+
+                    if (whitelists != null)
+                    {
+                        if (!whitelists.Exists(c => c.ChannelId == channel.Id))
+                        {
+                            await allWhitelists.InsertOneAsync(newWhite);
+                            await ReplyAsync($"All Commands are now Whitelisted to `{channel.Name}`");
+                            Logger.Info($"All Commands now whitelisted to {channel.Name} on {Context.Guild.Name}");
+                        }
+                        else
+                        {
+                            await ReplyAsync($"All Commands are already whitelisted to `{channel.Name}`");
+                            Logger.Warn($"Failed to Whitelist All Commands to {channel.Name} on {Context.Guild.Name}, already Whitelisted");
+                        }
                     }
                     else
                     {
-                        await ReplyAsync($"All Commands are already whitelisted to `{channel.Name}`");
-                        logger.Warn($"Failed to Whitelist All Commands to {channel.Name} on {Context.Guild.Name}, already Whitelisted");
+                        await allWhitelists.InsertOneAsync(newWhite);
+                        await ReplyAsync($"All Commands are now Whitelisted to `{channel.Name}`");
+                        Logger.Info($"All Commands are now whitelisted to {channel.Name} on {Context.Guild.Name}");
                     }
-                }
-                else
-                {
-                    await allWhitelists.InsertOneAsync(newWhite);
-                    await ReplyAsync($"All Commands are now Whitelisted to `{channel.Name}`");
-                    logger.Info($"All Commands are now whitelisted to {channel.Name} on {Context.Guild.Name}");
                 }
             }
 
@@ -148,33 +150,27 @@ namespace RubyRose.Modules.Moderation
 
                 if (whitelists != null)
                 {
-                    if (whitelists.Exists(c => c.ChannelId == channel.Id))
+                    if (whitelists.Exists(c => channel != null && c.ChannelId == channel.Id))
                     {
-                        var whitelist = whitelists.First(c => c.ChannelId == channel.Id);
+                        var whitelist = whitelists.First(c => channel != null && c.ChannelId == channel.Id);
                         await allWhitelists.DeleteAsync(whitelist);
-                        await ReplyAsync($"All Commands are no longer Whitelisted to `{channel.Name}`");
-                        logger.Info($"All Commands are no longer whitelisted to {channel.Name} on {Context.Guild.Name}");
+                        await ReplyAsync($"All Commands are no longer Whitelisted to `{channel?.Name}`");
+                        Logger.Info($"All Commands are no longer whitelisted to {channel?.Name} on {Context.Guild.Name}");
                     }
                     else
                     {
-                        await ReplyAsync($"No Command is not whitelisted to `{channel.Name}`");
-                        logger.Warn($"Failed to remove Whitelist for All Commands from {channel.Name} on {Context.Guild.Name}, not Whitelisted");
+                        await ReplyAsync($"No Command is not whitelisted to `{channel?.Name}`");
+                        Logger.Warn($"Failed to remove Whitelist for All Commands from {channel?.Name} on {Context.Guild.Name}, not Whitelisted");
                     }
                 }
                 else
                 {
-                    await ReplyAsync($"No Command is not whitelisted to `{channel.Name}`");
-                    logger.Warn($"Failed to remove Whitelist for All Commands from {channel.Name} on {Context.Guild.Name}, not Whitelisted");
+                    await ReplyAsync($"No Command is not whitelisted to `{channel?.Name}`");
+                    Logger.Warn($"Failed to remove Whitelist for All Commands from {channel?.Name} on {Context.Guild.Name}, not Whitelisted");
                 }
             }
 
-            private async Task<List<Whitelists>> GetWhitelistsAsync(IMongoCollection<Whitelists> collection, IGuild guild)
-            {
-                var whitelistsCursor = await collection.FindAsync(f => f.GuildId == guild.Id);
-                return await whitelistsCursor.ToListAsync();
-            }
-
-            private async Task<List<Whitelists>> GetCommandWhitelists(IMongoCollection<Whitelists> collection, IGuild guild, string name)
+            private static async Task<List<Whitelists>> GetCommandWhitelists(IMongoCollection<Whitelists> collection, IGuild guild, string name)
             {
                 var whitelistsCursor = await collection.FindAsync(f => f.GuildId == guild.Id && f.Name == name);
                 return await whitelistsCursor.ToListAsync();
